Move Specificators in the Tokening namespace
