+++ DubUrl.Core/Mapping/Tokening/ISpecificator.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DubUrl.Mapping
+namespace DubUrl.Mapping.Tokening
 {
     public interface ISpecificator
     {
