+++ DubUrl.Core/Mapping/Tokening/SpecificatorStraight.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace DubUrl.Mapping
+namespace DubUrl.Mapping.Tokening
 {
     internal class SpecificatorStraight : Specificator
     {
